Added auth blueprint.  Fixed errors BP.  Cleaned up login page and logic.
