+++ app/__init__.py
@@ -28,9 +28,13 @@ sshhandler = SSHHandler()
 celery = Celery(app.name, broker=app.config['CELERY_BROKER_URL'], backend=app.config['CELERY_RESULT_BACKEND'])
 celery.conf.update(app.config)
 
-# Blueprints
+# Errors blueprint
 from app.errors import bp as errors_bp
 app.register_blueprint(errors_bp)
+
+# Authentication blueprint
+from app.auth import bp as auth_bp
+app.register_blueprint(auth_bp, url_prefix='/auth')
 
 from app import views, models
