+++ app/templates/errors/notloggedin.html
@@ -1,6 +1,6 @@
 <h2><small>
-	Please set your AD credentials in the login page before proceeding.
+	Please set your SSH credentials in the login page before proceeding.
 </small></h2>
 <h4 class="text-primary">
-	<a href='/login'>Login</a>
+	<a href="{{ url_for('auth.login') }}">Login</a>
 </h4>
