+++ app/templates/inc/navbar-menu.html
@@ -65,10 +65,10 @@
               <a href="#" class="dropdown-toggle" data-toggle="dropdown" title="{{ session['USER'] }}" role="button" aria-haspopup="true" aria-expanded="false">{{ session['USER'] }} <b class="caret"></b></a>
               <ul class="dropdown-menu">
                 <li><a href="/disconnectAllSSH">Close Active SSH Sessions (<span id="activeSSHSessionCount"></span>)</a></li>
-                <li><a href="/logout">Logout</a></li>
+                <li><a href="/auth/logout">Logout</a></li>
               </ul>
             {% else %}
-              <a href="/login">Login</a>
+              <a href="/auth/login">Login</a>
             {% endif %}
           </li>
         </ul>
